Create README.md
This is a display of sample websites I have created with html and css.

1. space station

This page is an example homepage for a space trravel company. Its styled using css and contains a nav bar, image containers, and a form.


2. movie theater

This page is an example homepage for a move theater that is styled using Bootstrap. It contains many Bootstrap components including but not limited to a navbar, jumbotron, form, cards, and dropdowns.